move calculate_accuracy to BaseMixer
